refactor(Money.java): Money의 인스턴스 변수 타입을 double로 변경

+++ src/main/java/lotto/LottoGenerator.java
@@ -19,8 +19,8 @@ public class LottoGenerator {
     }
 
     public Lottos generate(Money money) {
-        int count = money.divide(price);
-        return generate(count);
+        double count = money.divide(price);
+        return generate((int) count);
     }
 
     private Lottos generate(int count) {
